+++ projects/smart-table/src/lib/smart-table/smart-table.types.ts
@@ -86,6 +86,7 @@ export enum SmartTableFilterType {
   Select = 'select',
   Input = 'input',
   Datepicker = 'datepicker',
+  SearchFilter = 'search-filter'
 }
 
 export enum SmartTableFilterDisplay {
@@ -96,7 +97,8 @@ export enum SmartTableFilterDisplay {
 
 export enum SmartTableFilterOperator {
   Equal = '=',
-  ILike = 'ILIKE'
+  ILike = 'ILIKE',
+  In = 'in'
 }
 
 export interface SmartTableFilterConfig {